refactor: organize tracker observability and aws helpers
