+++ services/tracker/main.py
@@ -22,15 +22,8 @@ from tracker.auth import (
     get_current_org,
     resolve_descope_tenant,
 )
-from tracker.cloudwatch import get_cloudwatch_url
-from tracker.config import AUTH_REQUIRED, ENVIRONMENT
-from tracker.database.models import Benchmark, BenchmarkStatus, Org, RetryMode
-from tracker.database.scoping import assert_org, get_scoped
-from tracker.database.session import check_database_connection, get_session
-from tracker.exceptions import TrackerServiceError
-from tracker.logging import benchmark_id_var, configure_logging, get_logger, request_id_var
-from tracker.middleware import RequestContextMiddleware
-from tracker.s3 import (
+from tracker.aws.cloudwatch_logs import get_benchmark_log_url
+from tracker.aws.s3 import (
     S3_BENCHMARKS_PREFIX,
     copy_agent_to_benchmark,
     create_benchmark_url,
@@ -40,8 +33,14 @@ from tracker.s3 import (
     list_s3_objects,
     s3_object_exists,
 )
-from tracker.sentry import init_sentry
-from tracker.tracing import configure_tracing
+from tracker.config import AUTH_REQUIRED, ENVIRONMENT
+from tracker.database.models import Benchmark, BenchmarkStatus, Org, RetryMode
+from tracker.database.scoping import assert_org, get_scoped
+from tracker.database.session import check_database_connection, get_session
+from tracker.exceptions import TrackerServiceError
+from tracker.logging import benchmark_id_var, configure_logging, get_logger, request_id_var
+from tracker.middleware import RequestContextMiddleware
+from tracker.observability import configure_observability
 from tracker.types import (
     BenchmarkTableRow,
     FetchBenchmarkMetadataResponse,
@@ -74,8 +73,7 @@ from tracker.utils import (
 )
 
 configure_logging()
-init_sentry("valkyrie-tracker", environment=ENVIRONMENT)
-configure_tracing("valkyrie-tracker", environment=ENVIRONMENT)
+configure_observability("valkyrie-tracker", environment=ENVIRONMENT)
 
 logger = get_logger(__name__)
 
@@ -260,7 +258,7 @@ async def start_benchmark(
         concurrency=request.concurrency,
         started_at=benchmark_row.started_at,
         task_count=len(verify_response.task_ids),
-        cloudwatch_url=get_cloudwatch_url(
+        cloudwatch_url=get_benchmark_log_url(
             str(benchmark_row.id), request.harness_config.aws.aws_default_region, request.harness_config.log_group
         ),
         s3_bucket_url=create_benchmark_url(
