+++ services/tracker/src/tracker/__init__.py
@@ -2,6 +2,6 @@
 Methods and types we want to export from the tracker to the CLI
 """
 
-from tracker.s3 import handle_s3_error
+from tracker.aws.s3 import handle_s3_error
 
 __all__ = ["handle_s3_error"]
