+++ services/tracker/src/tracker/aws/__init__.py
@@ -1,0 +1,2 @@
+"""AWS service helpers for the tracker service."""
+
