+++ services/tracker/src/tracker/aws/cloudwatch_logs.py
@@ -41,7 +41,7 @@ def handle_cloudwatch_error(message: str):
     return decorator
 
 
-def get_cloudwatch_url(benchmark_id: str, region: str, log_group: str, task_id: str | None = None) -> str:
+def get_benchmark_log_url(benchmark_id: str, region: str, log_group: str, task_id: str | None = None) -> str:
     """
     Get the CloudWatch console URL for a benchmark or specific task.
 
@@ -64,7 +64,9 @@ def get_cloudwatch_url(benchmark_id: str, region: str, log_group: str, task_id: 
 
 @handle_cloudwatch_error(message="Failed to create log group")
 @logfire.instrument("create_log_group", extract_args=("benchmark_id",))
-def create_benchmark_group(benchmark_id: str, aws: "AWSCredentials", log_group: str, log_retention_policy: int) -> str:
+def create_benchmark_log_group(
+    benchmark_id: str, aws: "AWSCredentials", log_group: str, log_retention_policy: int
+) -> str:
     """
     Create a log group for a benchmark.
 
@@ -91,7 +93,7 @@ def create_benchmark_group(benchmark_id: str, aws: "AWSCredentials", log_group: 
 
 
 @handle_cloudwatch_error(message="Failed to create cloudwatch stream")
-def cloudwatch_stream(stream_key: str, message: str, aws: "AWSCredentials", log_group: str) -> None:
+def write_benchmark_log_event(stream_key: str, message: str, aws: "AWSCredentials", log_group: str) -> None:
     """
     Stream a log message to CloudWatch.
