+++ services/tracker/src/tracker/config.py
@@ -10,8 +10,7 @@ from taskiq_redis.redis_backend import RedisAsyncResultBackend
 
 from tracker.logging import configure_logging
 from tracker.middleware import LoggingContextMiddleware, TaskProtectionMiddleware, TracingContextMiddleware
-from tracker.sentry import init_sentry
-from tracker.tracing import configure_tracing
+from tracker.observability import configure_observability
 
 load_dotenv()
 configure_logging()
@@ -74,8 +73,7 @@ broker = (
 
 @broker.on_event(TaskiqEvents.WORKER_STARTUP)
 async def _init_worker_observability(*_args: object, **_kwargs: object) -> None:  # pyright: ignore[reportUnusedFunction]
-    init_sentry("valkyrie-worker", environment=ENVIRONMENT)
-    configure_tracing("valkyrie-worker", environment=ENVIRONMENT)
+    configure_observability("valkyrie-worker", environment=ENVIRONMENT)
 
 
 # Auth settings
