+++ services/tracker/src/tracker/observability/sentry.py
@@ -2,6 +2,7 @@
 
 import logging
 import os
+from typing import Any
 
 import daytona
 import sentry_sdk
@@ -11,6 +12,9 @@ from sentry_sdk.types import Event, Hint
 
 from tracker.exceptions import PtyCreationError, SSLConnectionError, SandboxError
 from tracker.logging.context import get_context_tags
+from tracker.observability.metrics import incr
+
+logger = logging.getLogger(__name__)
 
 
 def _before_send(
@@ -71,4 +75,46 @@ def init_sentry(service_name: str, environment: str) -> None:
         sentry_sdk.set_tag("daytona.sdk_version", getattr(daytona, "__version__", "unknown"))
     except Exception as e:
         # A malformed SENTRY_DSN or invalid integration shouldn't crash service startup.
-        logging.getLogger(__name__).warning("Failed to initialize Sentry: %s: %s", type(e).__name__, e)
+        logger.warning("Failed to initialize Sentry: %s: %s", type(e).__name__, e)
+
+
+def set_sandbox_context(sandbox: Any, *, image: str | None = None) -> None:
+    """Attach sandbox identifiers to Sentry tags/context, not metric attributes."""
+    try:
+        sentry_sdk.set_tag("sandbox_id", sandbox.id)
+        sentry_sdk.set_tag("sandbox_name", sandbox.name)
+
+        state = getattr(sandbox, "state", None)
+        context = {
+            "id": sandbox.id,
+            "name": sandbox.name,
+            "state": str(state) if state is not None else None,
+        }
+        if image is not None:
+            context["image"] = image
+
+        sentry_sdk.set_context("sandbox", context)
+    except Exception as e:
+        logger.warning("set_sandbox_context failed: %s: %s", type(e).__name__, e)
+
+
+def set_pty_context(*, session_id: str, attempt: int | None = None) -> None:
+    """Attach PTY identifiers to Sentry tags, not metric attributes."""
+    try:
+        sentry_sdk.set_tag("pty_session_id", session_id)
+        if attempt is not None:
+            sentry_sdk.set_tag("pty_attempt", str(attempt))
+    except Exception as e:
+        logger.warning("set_pty_context failed: %s: %s", type(e).__name__, e)
+
+
+def tag_daytona_error(exc: Exception, *, op: str) -> None:
+    """Tag the active Sentry scope and emit a low-cardinality Daytona error metric."""
+    error_class = type(exc).__name__
+    try:
+        sentry_sdk.set_tag("daytona.op", op)
+        sentry_sdk.set_tag("error_class", error_class)
+    except Exception as e:
+        logger.warning("tag_daytona_error failed: %s: %s", type(e).__name__, e)
+
+    incr("valkyrie.daytona.error", tags={"op": op, "error_class": error_class})
