+++ services/tracker/src/tracker/observability/tracing.py
