+++ services/tracker/tests/integration/test_upload_artifacts_images.py
@@ -7,7 +7,7 @@ from daytona import AsyncDaytona
 
 from tests.utils import random_task_id
 from tracker.database.models import AgentContractRequest
-from tracker.s3 import get_benchmark_contract_s3_key, get_contract_s3_key
+from tracker.aws.s3 import get_benchmark_contract_s3_key, get_contract_s3_key
 from tracker.sandbox import create_sandbox, upload_agent_artifacts
 from tracker.types import AWSCredentials, HarnessConfig
