+++ services/tracker/tests/integration/test_upload_to_s3.py
@@ -4,7 +4,7 @@ from daytona import AsyncDaytona
 
 from tests.utils import random_task_id
 from tracker.database.models import Benchmark
-from tracker.s3 import delete_from_s3, download_from_s3, get_agent_result_s3_key
+from tracker.aws.s3 import delete_from_s3, download_from_s3, get_agent_result_s3_key
 from tracker.sandbox import TrackerResources, archive_and_upload_output, create_sandbox
 from tracker.types import HarnessConfig
